# specs/001-student-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Student Management System API
  version: 1.0.0
  description: REST API for student enrollment, viewing, and record management
  contact:
    name: Student Management System
    email: admin@schoolsystem.com
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/courses:
    get:
      summary: List all courses
      operationId: listCourses
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: teacherId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            description: Filter by teacher
        - name: gradeLevel
          in: query
          required: false
          schema:
            type: string
            maxLength: 20
            description: Filter by grade level label (free text, e.g. "9" or "K-1")
      responses:
        '200':
          description: Successful response with course list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CourseSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new course
      operationId: createCourse
      tags:
        - courses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Course successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/courses/{course_id}:
    get:
      summary: Get course by ID
      operationId: getCourse
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with course details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    put:
      summary: Update course information
      operationId: updateCourse
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Course successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/courses/{course_id}/students:
    get:
      summary: List students enrolled in a course
      operationId: listCourseStudents
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with enrolled students
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentSummary'
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    put:
      summary: Replace the roster of students assigned to a course
      operationId: updateCourseStudents
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseRosterUpdate'
      responses:
        '200':
          description: Course roster updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          description: Bad request - invalid or missing student IDs
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/courses/{course_id}/parents:
    get:
      summary: List parents of the students in a course (grouped per student)
      operationId: listCourseParents
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      description: >-
        Teachers can view this for their own courses so they can reach guardians
        about absences or poor performance; admins for any course. Non-own
        courses (teacher) and parents are forbidden.
      responses:
        '200':
          description: Successful response with per-student parent lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        student_id:
                          type: string
                          format: uuid
                        student_name:
                          type: string
                        parents:
                          type: array
                          items:
                            $ref: '#/components/schemas/ParentSummary'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not your course, or not a teacher/admin
  /api/v1/attendance:
    post:
      summary: Mark daily attendance for a course
      operationId: markAttendance
      tags:
        - attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceCreate'
      responses:
        '201':
          description: Attendance successfully recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
        '400':
          description: Bad request - validation failed (e.g., no students selected, invalid course)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - teacher not assigned to course
  /api/v1/attendance/{course_id}:
    get:
      summary: Get attendance records for a course
      operationId: listCourseAttendance
      tags:
        - attendance
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Filter attendance by date
      responses:
        '200':
          description: Successful response with attendance records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceDetail'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/students/{student_id}/attendance:
    get:
      summary: Get attendance records for a student
      operationId: getStudentAttendance
      tags:
        - attendance
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with student attendance records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceDetail'
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/export/students.csv:
    get:
      summary: CSV export of all students (admin-only)
      operationId: exportStudentsCsv
      tags:
        - export
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Comma-separated values (attachment). Formula-injection defused.
          content:
            text/csv:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not an administrator
  /api/v1/export/grades.csv:
    get:
      summary: CSV export of all grades with course + student context (admin-only)
      operationId: exportGradesCsv
      tags:
        - export
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Comma-separated values (attachment). Formula-injection defused.
          content:
            text/csv:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - not an administrator
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentSummary:
      type: object
      properties:
        student_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        grade_level:
          type: string
        active:
          type: boolean
        enrollment_date:
          type: string
          format: date
...
    CourseSummary:
      type: object
      properties:
        course_id:
          type: string
          format: uuid
        name:
          type: string
        teacher_id:
          type: string
          format: uuid
        grade_level:
          type: string
        semester:
          type: string
        status:
          type: string
    CourseDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/CourseSummary'
        - type: object
          properties:
            max_students:
              type: integer
            students:
              type: array
              items:
                $ref: '#/components/schemas/StudentSummary'
    CourseCreate:
      type: object
      required:
        - name
        - teacher_id
        - grade_level
        - semester
      properties:
        name:
          type: string
          maxLength: 100
        teacher_id:
          type: string
          format: uuid
        grade_level:
          type: string
          minLength: 1
          maxLength: 20
          description: School-defined grade label (e.g. "9", "K-1", "O-Level")
        semester:
          type: string
          maxLength: 20
        max_students:
          type: integer
          default: 30
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        teacher_id:
          type: string
          format: uuid
        max_students:
          type: integer
        status:
          type: string
          enum: [active, inactive, full]
    CourseResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        course:
          $ref: '#/components/schemas/CourseDetail'
    CourseRosterUpdate:
      type: object
      required:
        - student_ids
      properties:
        student_ids:
          type: array
          items:
            type: string
            format: uuid
    ParentSummary:
      type: object
      properties:
        parent_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: boolean
...
    AttendanceCreate:
      type: object
      required:
        - course_id
        - date
        - records
      properties:
        course_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        records:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - student_id
              - status
            properties:
              student_id:
                type: string
                format: uuid
              status:
                type: string
                enum: [present, absent, late, excused]
    AttendanceDetail:
      type: object
      properties:
        attendance_id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        marked_by:
          type: string
          format: uuid
    AttendanceResponse:
      type: object
      properties:
        success:
          type: boolean
        attendance:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceDetail'
        message:
          type: string
...